server/requirements.txt
Flask
openai
pandas
scikit-learn
numpy
python-dotenv
nltk